+++ src/app/pages/PerpetualPage/Loadable.tsx
@@ -1,9 +1,3 @@
-/**
- *
- * Asynchronously loads the component for MarginTradePage
- *
- */
-
 import React from 'react';
 import { lazyLoad } from 'utils/loadable';
 import { PageSkeleton } from 'app/components/PageSkeleton';
