+++ src/app/pages/PerpetualPage/components/CustomToastContent/index.tsx
@@ -29,26 +29,24 @@ export const CustomToastContent: React.FC<ICustomToastContentProps> = ({
   toastProps,
   mainInfo,
   additionalInfo,
-}) => {
-  return (
-    <div className="tw-flex tw-items-start">
-      <img
-        src={getIcon(toastProps?.type)}
-        alt="success"
-        className="tw-w-7 tw-mr-2.5"
-      />
+}) => (
+  <div className="tw-flex tw-items-start">
+    <img
+      src={getIcon(toastProps?.type)}
+      alt="success"
+      className="tw-w-7 tw-mr-2.5"
+    />
 
-      <div className="tw-flex tw-text-black tw-items-start">
-        <span className="tw-font-semibold tw-font-loose tw-mt-0.5">
-          {mainInfo}
-        </span>
-        <span className="tw-text-xs tw-font-normal tw-font-medium tw-mt-1.5 tw-ml-4">
-          {additionalInfo}
-        </span>
-      </div>
+    <div className="tw-flex tw-text-black tw-items-start">
+      <span className="tw-font-semibold tw-font-loose tw-mt-0.5">
+        {mainInfo}
+      </span>
+      <span className="tw-text-xs tw-font-normal tw-font-medium tw-mt-1.5 tw-ml-4">
+        {additionalInfo}
+      </span>
     </div>
-  );
-};
+  </div>
+);
 
 export const toastOptions: ToastOptions = {
   hideProgressBar: false,
