+++ src/app/pages/PerpetualPage/components/LeverageSelector/index.tsx
@@ -56,13 +56,16 @@ export const LeverageSelector: React.FC<LeverageSelectorProps> = ({
     setManual(true);
     onChange(min);
   }, [onChange, min]);
+
   const onDisableManual = useCallback(() => {
     setManual(false);
+
     const nearestStepValue = steps.reduce((nearest, step) => {
       const currentDifference = Math.abs(nearest - value);
       const stepDifference = Math.abs(step - value);
       return stepDifference <= currentDifference ? step : nearest;
     }, steps[0]);
+
     if (nearestStepValue !== value) {
       onChange(nearestStepValue);
     }
