+++ src/app/pages/PerpetualPage/components/PairSelector/index.tsx
@@ -1,16 +1,13 @@
-import React, { useContext, useMemo, useCallback } from 'react';
+import React, { useMemo } from 'react';
 import { PerpetualPair } from '../../../../../utils/models/perpetual-pair';
 import {
   PerpetualPairDictionary,
   PerpetualPairType,
 } from '../../../../../utils/dictionaries/perpetual-pair-dictionary';
-import classNames from 'classnames';
-import { getPriceColor, getPriceChange } from '../RecentTradesTable/utils';
-import { toNumberFormat } from '../../../../../utils/display-text/format';
-import { RecentTradesContext } from '../../contexts/RecentTradesContext';
 import { getCollateralName, getCollateralLogo } from '../../utils/renderUtils';
 import { Asset } from '../../../../../types';
 import { GsnSwitch } from '../GsnSwitch/GsnSwitch';
+import { PairSelectorButton } from './PairSelectorButton';
 
 type PairSelectorProps = {
   pair: PerpetualPair;
@@ -55,55 +52,6 @@ export const PairSelector: React.FC<PairSelectorProps> = ({
           <GsnSwitch />
         </div>
       </div>
-    </div>
-  );
-};
-
-type PairSelectorButtonProps = {
-  pair: PerpetualPair;
-  isSelected: boolean;
-  onSelect: (pair: PerpetualPairType) => void;
-};
-
-const PairSelectorButton: React.FC<PairSelectorButtonProps> = ({
-  pair,
-  isSelected,
-  onSelect,
-}) => {
-  const { trades } = useContext(RecentTradesContext);
-  const latestPrice = trades[0]?.price;
-  const previousPrice = trades[1]?.price;
-  const color = useMemo(
-    () =>
-      getPriceColor(getPriceChange(previousPrice || latestPrice, latestPrice)),
-    [previousPrice, latestPrice],
-  );
-
-  const onClick = useCallback(() => !isSelected && onSelect(pair.pairType), [
-    onSelect,
-    pair,
-    isSelected,
-  ]);
-
-  return (
-    <div
-      className={classNames(
-        'tw-flex tw-flex-row tw-items-center tw-min-w-56 tw-px-3 tw-py-1.5 tw-mr-2 tw-rounded-lg tw-select-none tw-transition-colors tw-duration-300',
-        isSelected
-          ? 'tw-bg-gray-5'
-          : 'tw-bg-gray-4 tw-cursor-pointer hover:tw-bg-gray-5',
-      )}
-      onClick={onClick}
-    >
-      <span className="tw-font-medium tw-mr-2 tw-text-xs">{pair.name}</span>
-      <span
-        className={classNames(
-          'tw-flex-auto tw-text-right tw-font-medium tw-text-base',
-          color,
-        )}
-      >
-        {toNumberFormat(latestPrice, 1)}
-      </span>
     </div>
   );
 };
