+++ src/app/pages/PerpetualPage/components/SlippageForm/index.tsx
@@ -16,6 +16,7 @@ const SLIPPAGE_MIN = 0.001;
 const SLIPPAGE_MAX = isMainnet ? 0.01 : 0.02;
 const SLIPPAGE_STEP = 0.0005;
 const SLIPPAGE_LABEL_COUNT = 5;
+
 const SLIPPAGE_LABELS = Array(SLIPPAGE_LABEL_COUNT)
   .fill(0)
   .map((_, index) =>
