+++ src/app/pages/PerpetualPage/components/TradeDialog/components/ResultPosition.tsx
@@ -33,6 +33,7 @@ export const ResultPosition: React.FC<ResultPositionProps> = ({
     limitPrice,
     liquidationPrice,
   } = analysis;
+
   const { t } = useTranslation();
 
   if (
